feat(config): add 'Zhi Mang Xing' font to Google Fonts (#130)
Adds the 'Zhi Mang Xing' font to the list of Google Fonts used in the
Nuxt.js application. This font is used in the custom Poe component to
display the poem text.

feat(module): format module date and display it

Formats the module date using the `dayjs` library and displays it in a
badge next to the module title. This provides more context about when
the module was published.

style(footer): adjust grid layout on smaller screens

Modifies the grid layout of the footer sitemap on smaller screens to
display the items in a two-column layout instead of four columns. This
improves the responsiveness and readability of the footer on mobile
devices.

feat(poe): increase font size and adjust animation

Increases the font size of the major and poe text in the custom Poe
component, making the text more prominent and readable. Additionally,
the animation duration of the poe text is adjusted to improve the
visual effect.

style(theme): adjust secondary color and foreground

Updates the secondary color and foreground in the Tailwind CSS
configuration to match the primary color, creating a more consistent
color scheme throughout the application.